requirements.txt
markdown==3.2.1
mkdocs==1.2.1
mkdocs-material==7.1.9
mkdocs-simple-hooks==0.1.3
pyyaml==5.4.1